Read a salt from a file in the user $HOME directory.
The salt come from the filesystem. Which means that if any 2 users
has the same password, they will be able to decrypt files of the other user
because they also share the same salt from the filesystem.
This allows each user to have their own salt.
The salt is read from the file $HOME/.ext4_encryption_salt and can contains a
salt of max 256 bytes in the following form:
  - an uuid
  - 0x prefix followed by the bytes of the salt in hexadecimal form
  - s: prefix followed by the bytes of the salt