Remove salt struct
The salt struct exists mostly for e4crypt's input handling and to
keep a single type across multiple subcommands. It's of no use to
pam_e4crypt so we get rid of it.

Since we're not parsing user input and not using a salt we'll also
get rid of salt_parse and just use the raw contents of the salt
file as a salt.