Add extra warnings for cases where we failed to generate a key
Salt retrievel relies on successful reads from files containing the
salt. This may fail for various reasons, the most common probably being
typos in a path or a salt-file which was not prepared properly. However,
we failed to report those most common cases to the user.

This commit introduces some warnings, hopefully helping users find
errors in their setup.